+++ test/chart/data_pool.dart
@@ -305,7 +305,7 @@ class MockData {
   );
 
   static final lineTouchResponse1 =
-      LineTouchResponse([lineBarSpot1, lineBarSpot2]);
+      LineTouchResponse([lineBarSpot1, lineBarSpot2], const FlSpot(1, 1));
 
   static final barChartRodData1 = BarChartRodData(toY: 11);
   static final barChartRodData2 = BarChartRodData(toY: 22);
@@ -1020,8 +1020,12 @@ final TouchLineBarSpot lineBarSpot1Clone = TouchLineBarSpot(
   0,
 );
 
-final TouchLineBarSpot lineBarSpot2 =
-    TouchLineBarSpot(lineChartBarData1, 2, flSpot1, 2);
+final TouchLineBarSpot lineBarSpot2 = TouchLineBarSpot(
+  lineChartBarData1,
+  2,
+  flSpot1,
+  2,
+);
 
 final TouchLineBarSpot lineBarSpot3 = TouchLineBarSpot(
   lineChartBarData1,
@@ -1035,12 +1039,14 @@ final LineTouchResponse lineTouchResponse1 = LineTouchResponse(
     lineBarSpot1,
     lineBarSpot2,
   ],
+  const FlSpot(1, 1),
 );
 final LineTouchResponse lineTouchResponse1Clone = LineTouchResponse(
   [
     lineBarSpot1Clone,
     lineBarSpot2,
   ],
+  const FlSpot(1, 1),
 );
 
 final LineTouchResponse lineTouchResponse2 = LineTouchResponse(
@@ -1048,17 +1054,18 @@ final LineTouchResponse lineTouchResponse2 = LineTouchResponse(
     lineBarSpot2,
     lineBarSpot1,
   ],
+  const FlSpot(1, 1),
 );
 
-const LineTouchResponse lineTouchResponse3 = LineTouchResponse(
-  [],
-);
+const LineTouchResponse lineTouchResponse3 =
+    LineTouchResponse([], FlSpot(1, 1));
 
 final LineTouchResponse lineTouchResponse4 = LineTouchResponse(
   [
     lineBarSpot1,
     lineBarSpot2,
   ],
+  const FlSpot(1, 1),
 );
 
 final LineTouchResponse lineTouchResponse5 = LineTouchResponse(
@@ -1066,6 +1073,7 @@ final LineTouchResponse lineTouchResponse5 = LineTouchResponse(
     lineBarSpot1,
     lineBarSpot2,
   ],
+  const FlSpot(1, 1),
 );
 
 const TouchedSpotIndicatorData touchedSpotIndicatorData1 =